make our own Proxy-State 64-bit
which means that collisions are essentially impossible.

+++ src/protocols/radius/client.h
@@ -44,7 +44,7 @@ typedef struct {
 	fr_time_delta_t		connection_timeout;
 
 	bool			add_proxy_state;
-	uint32_t		proxy_state;
+	uint64_t		proxy_state;
 
 	bool			outgoing[FR_RADIUS_CODE_MAX];	//!< allowed outgoing packet types
